image_processor/requirements.txt
asn1crypto==0.23.0
azure-common==1.1.8
azure-nspkg==2.0.0
azure-storage-common==0.37.1
azure-storage-nspkg==2.0.0
azure-storage-queue==0.37.0
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
click==6.7
cryptography==2.1.4
dlib==19.7.0
face-recognition==1.0.0
face-recognition-models==0.3.0
futures==3.1.1
idna==2.6
imageio==2.2.0
mysql-connector-python==8.0.5
numpy==1.13.3
olefile==0.44
Pillow==4.3.0
pycparser==2.18
python-dateutil==2.6.1
requests==2.18.4
scipy==1.0.0
six==1.11.0
SQLAlchemy==1.1.15
urllib3==1.22

# pip install mysql-connector-python
# pip install sqlalchemy
# pip install Pillow
# pip install imageio
# pip install azure-storage-queue
# pip install numpy
# pip install scipy
# pip install face_recognition
